Create utils::test module for generators and collectors
There might be other non-test-related utilities we might want to add
under 'utils', so I created the submodule 'utils::test' to house packet
generators, packet collectors, and eventually some default testing
Elements or configurations.

Arguably, IdentityElement and AsyncIdentityElement belong in a
'utils::test::elements' module.
