+++ src/lib.rs
@@ -10,7 +10,8 @@ mod utils;
 mod tests {
     use super::*;
     use crate::api::{ElementLink, Element, AsyncElementLink, AsyncElement};
-    use crate::utils::{LinearIntervalGenerator, ExhaustiveDrain};
+    use crate::utils::test::packet_generators::LinearIntervalGenerator;
+    use crate::utils::test::packet_collectors::ExhaustiveDrain;
     use core::time;
     use futures::stream::iter_ok;
     use futures::future::lazy;
