+++ src/utils/test/packet_collectors.rs
@@ -1,0 +1,38 @@
+use crate::api::ElementStream;
+use futures::{Async, Poll, Future};
+use std::fmt::Debug;
+
+#[allow(dead_code)]
+pub struct ExhaustiveDrain<T: Debug> {
+    id: usize,
+    stream: ElementStream<T>
+}
+
+#[allow(dead_code)]
+impl<T: Debug> ExhaustiveDrain<T> {
+    pub fn new(id: usize, stream: ElementStream<T>) -> Self {
+        ExhaustiveDrain { id, stream }
+    }
+}
+
+#[allow(dead_code)]
+impl<T: Debug> Future for ExhaustiveDrain<T> {
+    type Item = ();
+    type Error = ();
+
+    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
+        // println!("Drain #{} poll", self.id);
+
+        loop {
+            match try_ready!(self.stream.poll()) {
+                Some(value) => {
+                    println!("Drain #{} received packet: {:?}", self.id, value);
+                },
+                None => {
+                    println!("Drain #{} received none. End of packet stream", self.id);
+                    return Ok(Async::Ready(()))
+                }
+            }
+        }
+    }
+}
